--- packages/libs/react-ui/src/components/NewTabs/index.ts
@@ -1,1 +1,0 @@
-export { TabItem, Tabs } from './Tabs';
