+++ packages/libs/react-ui/src/components/Tabs/TabPanel.tsx
@@ -3,6 +3,7 @@ import React, { useRef } from 'react';
 import type { AriaTabPanelProps } from 'react-aria';
 import { useTabPanel } from 'react-aria';
 import type { TabListState } from 'react-stately';
+import { tabContentClass } from './Tabs.css';
 
 interface ITabPanelProps extends AriaTabPanelProps {
   state: TabListState<object>;
@@ -14,8 +15,9 @@ interface ITabPanelProps extends AriaTabPanelProps {
 export const TabPanel = ({ state, ...props }: ITabPanelProps): ReactNode => {
   const ref = useRef(null);
   const { tabPanelProps } = useTabPanel(props, state, ref);
+
   return (
-    <div {...tabPanelProps} ref={ref}>
+    <div className={tabContentClass} {...tabPanelProps} ref={ref}>
       {state.selectedItem?.props.children}
     </div>
   );
