+++ packages/libs/react-ui/src/components/index.ts
@@ -69,7 +69,7 @@ export type {
   IThProps,
   ITrProps,
 } from './Table';
-export type { ITabContentProps, ITabProps, ITabsContainerProps } from './Tabs';
+export type { ITabsProps } from './Tabs';
 export type { ITagProps } from './Tag';
 export type { ITooltipProps } from './Tooltip';
 export type { ILabelValue, ITrackerCardProps } from './TrackerCard/TrackerCard';
@@ -124,7 +124,7 @@ export { Pagination } from './Pagination';
 export { ProfileSummary } from './ProfileSummary';
 export { ProgressBar } from './ProgressBar';
 export { Table } from './Table';
-export { Tabs } from './Tabs';
+export { TabItem, Tabs } from './Tabs';
 export { Tag } from './Tag';
 export { Tooltip } from './Tooltip';
 export { TrackerCard } from './TrackerCard/TrackerCard';
