+++ packages/libs/react-ui/src/index.ts
@@ -57,6 +57,7 @@ export type {
   ITabProps,
   ITableProps,
   ITabsContainerProps,
+  ITabsProps,
   ITagProps,
   ITdProps,
   ITextFieldProps,
@@ -108,6 +109,7 @@ export {
   SelectField,
   Stack,
   SystemIcon,
+  TabItem,
   Table,
   Tabs,
   Tag,
